+++ backend/api/src/main/java/ru/naumen/ectmapi/controllers/CommentController.java
@@ -2,18 +2,17 @@ package ru.naumen.ectmapi.controllers;
 
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 import ru.naumen.ectmapi.converter.CommentConverter;
-import ru.naumen.ectmapi.dto.TreeDto;
-import ru.naumen.ectmapi.dto.СommentDto;
-import ru.naumen.ectmapi.entity.Comment;
+import ru.naumen.ectmapi.dto.CommentDto;
 import ru.naumen.ectmapi.service.CommentService;
 
 
 @RestController
-@RequiredArgsConstructor
+@RequiredArgsConstructor(onConstructor_ = {@Autowired})
 @RequestMapping(value = "api/comment", produces = MediaType.APPLICATION_JSON_VALUE)
 public class CommentController {
 
@@ -23,14 +22,14 @@ public class CommentController {
     @Operation(summary = "Сохраняет комментарий")
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping("/save")
-    public void save(@RequestBody СommentDto commentDto){
+    public void save(@RequestBody CommentDto commentDto){
         commentService.save(commentConverter.fromDto(commentDto));
     }
   
     @Operation(summary = "Предоставляет комментарий по id")
     @ResponseStatus(HttpStatus.OK)
     @GetMapping("/get/{id}")
-    public СommentDto get(@PathVariable Long id) {
+    public CommentDto get(@PathVariable Long id) {
         return commentConverter.toDto(commentService.get(id));
     }
