+++ backend/api/src/main/java/ru/naumen/ectmapi/converter/CommentConverter.java
@@ -2,12 +2,12 @@ package ru.naumen.ectmapi.converter;
 
 import org.mapstruct.Mapper;
 import ru.naumen.ectmapi.entity.Comment;
-import ru.naumen.ectmapi.dto.СommentDto;
+import ru.naumen.ectmapi.dto.CommentDto;
 
 @Mapper(componentModel = "spring")
 public interface CommentConverter {
 
-    СommentDto toDto(Comment comment);
+    CommentDto toDto(Comment comment);
 
-    Comment fromDto(СommentDto commentDto);
+    Comment fromDto(CommentDto commentDto);
 }
