+++ backend/api/src/main/java/ru/naumen/ectmapi/service/CommentService.java
@@ -1,12 +1,13 @@
 package ru.naumen.ectmapi.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.naumen.ectmapi.entity.Comment;
 import ru.naumen.ectmapi.repository.CommentRepository;
 
 @Service
-@RequiredArgsConstructor
+@RequiredArgsConstructor(onConstructor_ = {@Autowired})
 public class CommentService {
 
     private final CommentRepository commentRepository;
@@ -16,7 +17,7 @@ public class CommentService {
     }
 
     public Comment get(Long id){
-        return commentRepository.findById(id).get();
+        return commentRepository.findById(id).orElseThrow();
     }
 
     public void delete(Long id) {commentRepository.deleteById(id);}
