+++ backend/api/src/main/java/ru/naumen/ectmapi/service/TreeService.java
@@ -1,12 +1,13 @@
 package ru.naumen.ectmapi.service;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.naumen.ectmapi.entity.Tree;
 import ru.naumen.ectmapi.repository.TreeRepository;
 
 @Service
-@RequiredArgsConstructor
+@RequiredArgsConstructor(onConstructor_ = {@Autowired})
 public class TreeService {
 
     private final TreeRepository treeRepository;
@@ -16,7 +17,7 @@ public class TreeService {
     }
 
     public Tree get(Long id){
-        return treeRepository.findById(id).get();
+        return treeRepository.findById(id).orElseThrow();
     }
 
     public void delete(Long id) {treeRepository.deleteById(id);}
